Convert and update
Added comments and made sure everything compiled in quincy
Quincy does not work with Lambda functions and had to make a second function call for it to work right.
Also had  to make sure pow had  a (double,int) by casting and using '10.0' over '10'.
Shortened C3_P3 display to fit a Windows 7 console
Also can not initialize a vector with '={}' in quincy, so used '(7)' instead, which won't give them a default value, but isn't needed for this program.

+++ CS102_Projects/C3/Marcum_C3_P17.cpp
@@ -70,8 +70,10 @@ int tutor(int dif){
   int num1, num2, ops = rand()%dif + 1;
   double ans = 0;
   //set numbers = to a random number with respect to a power of the difficulty
-  num1 = rand()%static_cast<int>(pow(10,dif)) + 1;
-  num2 = rand()%static_cast<int>(pow(10,dif)) + 1;
+  //Had to static_cast (dif) to get it to compile in Quincy, had to use 10.0 instead of 10 because of overload ambiguity
+  //static_cast on the result because of modulus only working with int values
+  num1 = rand()%static_cast<int>(pow(10.0,static_cast<int>(dif))) + 1;
+  num2 = rand()%static_cast<int>(pow(10.0,static_cast<int>(dif))) + 1;
   //output num1
   cout<< setw(10) << num1 << endl;
   //determine which operation will take place, priint it and calculate the answer
@@ -146,3 +148,4 @@ void s_i_in(int *input, int check[], int array_size){
     }
     *input = temp;
 }
+
